requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
discord.py==1.2.3
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauthlib==3.0.2
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
